fix(DataStore): Store larger than 32-bit values in Int64 over Int (#3367)
* fix(DataStore): Store larger than 32-bit values in Int64 over Int

* fix: Removing ModelFieldType.int64

+++ AmplifyPlugins/Core/AWSPluginsCore/Model/Decorator/ConflictResolutionDecorator.swift
@@ -15,12 +15,12 @@ import Foundation
 public struct ConflictResolutionDecorator: ModelBasedGraphQLDocumentDecorator {
 
     private let version: Int?
-    private let lastSync: Int?
+    private let lastSync: Int64?
     private let graphQLType: GraphQLOperationType
     private var primaryKeysOnly: Bool
     
     public init(version: Int? = nil,
-                lastSync: Int? = nil,
+                lastSync: Int64? = nil,
                 graphQLType: GraphQLOperationType,
                 primaryKeysOnly: Bool = true) {
         self.version = version
